Fix header using server code
